desktop_app/requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
mediapipe>=0.10.0
face_recognition>=1.3.0
SQLAlchemy>=2.0.0
edge-tts>=6.1.0
reportlab>=4.0.0
rapidfuzz>=3.0.0
paho-mqtt>=1.6.0
pyserial>=3.5
cryptography>=41.0.0
Pillow>=10.0.0
numpy>=1.24.0
tkcalendar>=1.6.0
pygame>=2.5.0
qrcode[pil]>=7.0

